Major refactoring of boto3-refresh-session
The primary purpose of this work is to reorganize boto3-refresh-session such that users are only exposed to the RefreshableSession object in session.py. All supporting internal objects are moved to utils, such as BaseRefreshableSession, et al.
