+++ .flake8
@@ -1,4 +1,4 @@
 [flake8]
 max-line-length = 79
 per-file-ignores = 
-    __init__.py:F401
+    */__init__.py:F401,F403
