+++ boto3_refresh_session/__init__.py
@@ -1,9 +1,14 @@
-from .methods.custom import CustomRefreshableSession
-from .methods.ecs import ECSRefreshableSession
-from .methods.sts import STSRefreshableSession
-from .session import RefreshableSession
+__all__ = []
 
-__all__ = ["RefreshableSession"]
+from . import exceptions, session
+from .exceptions import *
+from .methods.custom import *
+from .methods.ecs import *
+from .methods.sts import *
+from .session import *
+
+__all__.extend(session.__all__)
+__all__.extend(exceptions.__all__)
 __version__ = "2.0.11"
 __title__ = "boto3-refresh-session"
 __author__ = "Mike Letts"
