+++ boto3_refresh_session/methods/__init__.py
@@ -1,0 +1,12 @@
+__all__ = []
+
+# TODO: import iot submodules when finished
+from . import custom, ecs, sts
+from .custom import CustomRefreshableSession
+from .ecs import ECSRefreshableSession
+from .sts import STSRefreshableSession
+
+# TODO: add iot submodules to __all__ when finished
+__all__.extend(custom.__all__)
+__all__.extend(ecs.__all__)
+__all__.extend(sts.__all__)
